% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scottstock.R
\name{scottstock}
\alias{scottstock}
\title{scottstock}
\usage{
scottstock(ticker, frequency, start_date, end_date)
}
\arguments{
\item{ticker}{The stock symbol you want: "MSFT", or you can specify multiple by c("MSFT", "FB")}

\item{frequency}{YourThe type of data you want to pull: "daily", "weekly", "monthly", "yearly"}

\item{start_date}{The earliest date from which you want to pull data: "MM-DD-YYYY" surrounded by quotes}

\item{end_date}{The latest date from which you want to pull data: "MM-DD-YYYY" surrounded by quotes}
}
\value{
Output is a dataframe with the stock data specified.
\code{scottstock()} returns the Yahoo Finance data for a given ticker and start/end dates.
}
\description{
Yahoo Finance data analysis in R
}
\examples{
myStock <- scottstock("MSFT","monthly", "01-01-2019", "12-31-2019")
}
